moved jsx files into structure recommended by @timmyc
