+++ client/my-sites/stats/stats-detail-months/index.jsx
