+++ client/my-sites/stats/stats-detail-weeks/index.jsx
